MAINT: fix pandas warning re default dtype

+++ empyrical/utils.py
@@ -186,7 +186,7 @@ def _roll_pandas(func, window, *args, **kwargs):
         index_value = args[0].index[i - 1]
         index_values.append(index_value)
         data[index_value] = func(*rets, **kwargs)
-    return pd.Series(data, index=type(args[0].index)(index_values))
+    return pd.Series(data, index=type(args[0].index)(index_values), dtype="float64")
 
 
 @deprecated(msg=DATAREADER_DEPRECATION_WARNING)
